+++ test_project/views.py
@@ -3,7 +3,7 @@
 from rest_framework.viewsets import ModelViewSet
 
 from django_auto_prefetching import AutoPrefetchViewSetMixin
-from test_project.many_to_many_serializer import ManyTwoSerializerOnlyFullRepresentation
+from test_project.serializers.many_to_many_serializer import ManyTwoSerializerOnlyFullRepresentation
 from test_project.models import ManyToManyModelTwo
